# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListingImageRead": {
        "properties": {
          "blur_data_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blur Data Url"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "display_order": {
            "title": "Display Order",
            "type": "integer"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "r2_url": {
            "title": "R2 Url",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "r2_url",
          "display_order"
        ],
        "title": "ListingImageRead",
        "type": "object"
      },
      "ListingRead": {
        "properties": {
          "agent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/app__features__listings__schemas__AgentRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "amenities": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenities"
          },
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "bathroom_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bathroom Type"
          },
          "campus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campus Id"
          },
          "cooking_gas_included": {
            "default": false,
            "title": "Cooking Gas Included",
            "type": "boolean"
          },
          "county": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "County"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "distance_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Category"
          },
          "distance_to_campus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance To Campus"
          },
          "electricity_included": {
            "default": false,
            "title": "Electricity Included",
            "type": "boolean"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "hot_water_included": {
            "default": false,
            "title": "Hot Water Included",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "images": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ListingImageRead"
            },
            "title": "Images",
            "type": "array"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "is_full": {
            "default": false,
            "title": "Is Full",
            "type": "boolean"
          },
          "is_saved": {
            "default": false,
            "title": "Is Saved",
            "type": "boolean"
          },
          "is_youtube_shorts": {
            "default": false,
            "title": "Is Youtube Shorts",
            "type": "boolean"
          },
          "landlord_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Landlord Phone"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "pays_commission": {
            "default": false,
            "title": "Pays Commission",
            "type": "boolean"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "price_sharing": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Sharing"
          },
          "price_single": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Single"
          },
          "rating": {
            "default": 0.0,
            "title": "Rating",
            "type": "number"
          },
          "room_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Type"
          },
          "room_types": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ListingRoomTypeRead"
            },
            "title": "Room Types",
            "type": "array"
          },
          "security_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Security Type"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "sort_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Position"
          },
          "specific_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specific Location"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "views": {
            "default": 0,
            "title": "Views",
            "type": "integer"
          },
          "water_included": {
            "default": false,
            "title": "Water Included",
            "type": "boolean"
          },
          "wifi_included": {
            "default": false,
            "title": "Wifi Included",
            "type": "boolean"
          },
          "youtube_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube Id"
          },
          "zone_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone Id"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "price",
          "location",
          "created_at"
        ],
        "title": "ListingRead",
        "type": "object"
      },
...
      "PaginatedResponse_ListingRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ListingRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "example": 20,
            "title": "Limit",
            "type": "integer"
          },
          "page": {
            "example": 1,
            "title": "Page",
            "type": "integer"
          },
          "pages": {
            "example": 5,
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "example": 100,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "pages"
        ],
        "title": "PaginatedResponse[ListingRead]",
        "type": "object"
      },
...
      "ProfileRead": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "campus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campus Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "home_campus_confirmed": {
            "default": false,
            "title": "Home Campus Confirmed",
            "type": "boolean"
          },
          "home_campus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Campus Id"
          },
          "home_campus_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Campus Name"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "managed_campus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Managed Campus Id"
          },
          "managed_region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Managed Region Id"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "default": "student",
            "title": "Role",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "created_at"
        ],
        "title": "ProfileRead",
        "type": "object"
      },
      "ProfileUpdate": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "home_campus_confirmed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Campus Confirmed"
          },
          "home_campus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Campus Id"
          },
          "home_campus_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Campus Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "SavedHostelActionResponse": {
        "properties": {
          "is_saved": {
            "title": "Is Saved",
            "type": "boolean"
          },
          "listing_id": {
            "title": "Listing Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message",
          "is_saved",
          "listing_id"
        ],
        "title": "SavedHostelActionResponse",
        "type": "object"
      },
      "SetHomeCampusRequest": {
        "properties": {
          "campus_id": {
            "title": "Campus Id",
            "type": "string"
          },
          "campus_name": {
            "title": "Campus Name",
            "type": "string"
          }
        },
        "required": [
          "campus_id",
          "campus_name"
        ],
        "title": "SetHomeCampusRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__features__listings__schemas__AgentRead": {
        "properties": {
          "campus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campus Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "whatsapp": {
            "title": "Whatsapp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "whatsapp",
          "status"
        ],
        "title": "AgentRead",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Rumia Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/profiles/me": {
      "get": {
        "description": "Fetch profile of currently authenticated user. Authenticated.",
        "operationId": "get_my_profile_api_v1_profiles_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Current User Profile",
        "tags": [
          "Profiles"
        ]
      },
      "patch": {
        "description": "Update profile preferences. Authenticated.",
        "operationId": "update_my_profile_api_v1_profiles_me_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/v1/profiles/me/campus": {
      "post": {
        "description": "Set primary student campus. Authenticated.",
        "operationId": "set_home_campus_api_v1_profiles_me_campus_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetHomeCampusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Home Campus",
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/v1/profiles/me/saved": {
      "get": {
        "description": "Retrieve user's saved hostels. Authenticated.",
        "operationId": "get_saved_hostels_api_v1_profiles_me_saved_get",
        "parameters": [
          {
            "description": "Page number (1-indexed)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number (1-indexed)",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page (max 1000)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page (max 1000)",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ListingRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Saved Hostels",
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/v1/profiles/me/saved/{listing_id}": {
      "delete": {
        "description": "Remove a hostel from current user's favorites. Authenticated.",
        "operationId": "unsave_hostel_api_v1_profiles_me_saved__listing_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "title": "Listing Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedHostelActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unsave Hostel",
        "tags": [
          "Profiles"
        ]
      },
      "get": {
        "description": "Check whether the current user has saved a hostel. Authenticated.",
        "operationId": "get_saved_state_api_v1_profiles_me_saved__listing_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "title": "Listing Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedHostelActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Saved State",
        "tags": [
          "Profiles"
        ]
      },
      "post": {
        "description": "Save a hostel to current user's favorites. Authenticated.",
        "operationId": "save_hostel_api_v1_profiles_me_saved__listing_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "title": "Listing Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedHostelActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Hostel",
        "tags": [
          "Profiles"
        ]
      }
    },
...
  }
}